修复问题

+++ parent/src/main/java/fsy/utils/ReducePhotoUtils.java
@@ -23,16 +23,16 @@ public class ReducePhotoUtils {
     private Image img;
     private int width;
     private int height;
-    private String  originalFilename;
+    private String originalFilename;
 
     @SuppressWarnings("deprecation")
     public static void main(String[] args) throws Exception {
         System.out.println("开始：" + new Date().toLocaleString());
-        ReducePhotoUtils imgCom = new ReducePhotoUtils("C:\\Users\\admin\\Desktop\\项目流程图(1).png","项目流程图(1).png");
+        ReducePhotoUtils imgCom = new ReducePhotoUtils("C:\\Users\\admin\\Desktop\\项目流程图(1).png", "项目流程图(1).png");
         System.out.println("结束：" + new Date().toLocaleString());
     }
 
-    public ReducePhotoUtils(String fileName,String originalFilename) throws IOException {
+    public ReducePhotoUtils(String fileName, String originalFilename) throws IOException {
         this.originalFilename = originalFilename;
         File file = new File(fileName);// 读入文件
         img = ImageIO.read(file);      // 构造Image对象
@@ -43,6 +43,7 @@ public class ReducePhotoUtils {
 
     /**
      * 按照宽度还是高度进行压缩
+     *
      * @param w int 最大宽度
      * @param h int 最大高度
      */
@@ -53,37 +54,45 @@ public class ReducePhotoUtils {
             resizeByHeight(h);
         }
     }
+
     /**
      * 以宽度为基准，等比例放缩图片
+     *
      * @param w int 新宽度
      */
     public void resizeByWidth(int w) throws IOException {
         int h = (int) (height * w / width);
         resize(w, h);
     }
+
     /**
      * 以高度为基准，等比例缩放图片
+     *
      * @param h int 新高度
      */
     public void resizeByHeight(int h) throws IOException {
         int w = (int) (width * h / height);
         resize(w, h);
     }
+
     /**
      * 强制压缩/放大图片到固定的大小
+     *
      * @param w int 新宽度
      * @param h int 新高度
      */
     public void resize(int w, int h) throws IOException {
         // SCALE_SMOOTH 的缩略算法 生成缩略图片的平滑度的 优先级比速度高 生成的图片质量比较好 但速度慢
-        BufferedImage image = new BufferedImage(w, h,BufferedImage.TYPE_INT_RGB );
+        BufferedImage image = new BufferedImage(w, h, BufferedImage.TYPE_INT_RGB);
         image.getGraphics().drawImage(img, 0, 0, w, h, null); // 绘制缩小后的图
-        File destFile = new File(Const.WINDOWS_FILE_PATH+"min-"+this.originalFilename);
-        FileOutputStream out = new FileOutputStream(destFile); // 输出到文件流
-        // 可以正常实现bmp、png、gif转jpg
-        JPEGImageEncoder encoder = JPEGCodec.createJPEGEncoder(out);
-        encoder.encode(image); // JPEG编码
-        out.close();
+        File destFile = new File(("linux".equals(Const.OSType()) ? Const.LINUX_FILE_PATH : Const.WINDOWS_FILE_PATH) + Const.OSSeparator()+"min-" + this.originalFilename);
+        if (!destFile.exists()) {
+            FileOutputStream out = new FileOutputStream(destFile); // 输出到文件流
+            // 可以正常实现bmp、png、gif转jpg
+            JPEGImageEncoder encoder = JPEGCodec.createJPEGEncoder(out);
+            encoder.encode(image); // JPEG编码
+            out.close();
+        }
     }
 }
